moved animation funcs workhorse -> transforms
